+++ src/extension/Cycode.VisualStudio.Extension.Shared/Commands/OpenToolWindowCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Cycode.VisualStudio.Extension.Shared.Components.ToolWindows;
 
-namespace Cycode.VisualStudio.Extension.Shared;
+namespace Cycode.VisualStudio.Extension.Shared.Commands;
 
 [Command(PackageIds.ViewOpenToolWindowCommand)]
 internal sealed class ViewOpenToolWindowCommand : BaseCommand<ViewOpenToolWindowCommand> {