+++ src/extension/Cycode.VisualStudio.Extension.Shared/Components/ViolationCards/Common/TextBoxes/FieldValue.xaml.cs
@@ -1,4 +1,4 @@
-﻿namespace Cycode.VisualStudio.Extension.Shared.Components.ViolationCards;
+﻿namespace Cycode.VisualStudio.Extension.Shared.Components.ViolationCards.Common.TextBoxes;
 
 public partial class FieldValue {
     public FieldValue() {