+++ src/extension/Cycode.VisualStudio.Extension.Shared/CycodePackage.cs
@@ -5,9 +5,11 @@ global using Task = System.Threading.Tasks.Task;
 using System.Runtime.InteropServices;
 using System.Threading;
 using Cycode.VisualStudio.Extension.Shared.Components.ToolWindows;
+using Cycode.VisualStudio.Extension.Shared.Options;
 using Cycode.VisualStudio.Extension.Shared.Sentry;
 using Cycode.VisualStudio.Extension.Shared.Services;
 using Cycode.VisualStudio.Extension.Shared.Services.ErrorList;
+using Cycode.VisualStudio.Extension.Shared.Services.ErrorTagger;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Cycode.VisualStudio.Extension.Shared;