+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/CliDownloadService.cs
@@ -2,8 +2,13 @@ using System.IO;
 using System.Threading.Tasks;
 using Cycode.VisualStudio.Extension.Shared.DTO;
 using Cycode.VisualStudio.Extension.Shared.Helpers;
+using Cycode.VisualStudio.Extension.Shared.Options;
 
 namespace Cycode.VisualStudio.Extension.Shared.Services;
+
+public interface ICliDownloadService {
+    Task<bool> InitCliAsync();
+}
 
 public class CliDownloadService(
     ILoggerService logger,