+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/ICliDownloadService.cs
@@ -1,7 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Cycode.VisualStudio.Extension.Shared.Services;
-
-public interface ICliDownloadService {
-    Task<bool> InitCliAsync();
-}