+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/ScanResultsService.cs
@@ -6,6 +6,18 @@ using Cycode.VisualStudio.Extension.Shared.Cli.DTO.ScanResult.Secret;
 
 namespace Cycode.VisualStudio.Extension.Shared.Services;
 
+public interface IScanResultsService {
+    void SetSecretResults(SecretScanResult result);
+    SecretScanResult GetSecretResults();
+    void SetScaResults(ScaScanResult result);
+    ScaScanResult GetScaResults();
+    void Clear();
+    bool HasResults();
+    void SaveDetectedSegment(CliScanType scanType, TextRange textRange, string value);
+    string GetDetectedSegment(CliScanType scanType, TextRange textRange);
+    void ExcludeResultsByValue(string value);
+}
+
 public class ScanResultsService : IScanResultsService {
     private readonly Dictionary<(CliScanType, TextRange), string> _detectedSegments = new();
     private ScaScanResult _scaResults;
@@ -47,6 +59,11 @@ public class ScanResultsService : IScanResultsService {
     public string GetDetectedSegment(CliScanType scanType, TextRange textRange) {
         _detectedSegments.TryGetValue((scanType, textRange), out string value);
         return value;
+    }
+
+    public void ExcludeResultsByValue(string value) {
+        // we have value only in secret results
+        _secretResults?.Detections.RemoveAll(detection => detection.DetectionDetails.DetectedValue == value);
     }
 
     private void ClearDetectedSegments(CliScanType? scanType = null) {