+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/SuggestedActions/Actions/OpenViolationCardAction.cs
@@ -1,23 +1,21 @@
 ﻿using Cycode.VisualStudio.Extension.Shared.Cli.DTO;
 using Cycode.VisualStudio.Extension.Shared.Components.ToolWindows;
 using Cycode.VisualStudio.Extension.Shared.DTO;
-using Cycode.VisualStudio.Extension.Shared.Services.ErrorList;
+using Cycode.VisualStudio.Extension.Shared.Services.ErrorTagger;
 
 namespace Cycode.VisualStudio.Extension.Shared.Services.SuggestedActions.Actions;
 
 public class OpenViolationCardAction(DetectionTag tag) : BaseAction {
-    private readonly ILoggerService _logger = ServiceLocator.GetService<ILoggerService>();
-
     private static readonly IToolWindowMessengerService _toolWindowMessengerService =
         ServiceLocator.GetService<IToolWindowMessengerService>();
+
+    private readonly ILoggerService _logger = ServiceLocator.GetService<ILoggerService>();
 
     public override string DisplayText {
         get {
             string text = Tag.Detection.GetFormattedMessage();
             // cut too long messages
-            if (text.Length > 50) {
-                text = text.Substring(0, 50) + "...";
-            }
+            if (text.Length > 50) text = text.Substring(0, 50) + "...";
 
             return $"Cycode: {text}";
         }